Ssprod 5727 prevent concurrent policies load (#2329)
* Don't allow multiple policies load at once

If the future created by the std::async() call in
security_mgr::load_policies_v2_async() is valid, it means that a load is
in progress or is waiting to be picked up by process_event_v2(). In
either case, skip the (redundant) load of policies.

This fixes https://sysdig.atlassian.net/browse/SSPROD-5727.

* Add test for load/reload in parallel

Add a test that repeatedly loads + reloads the set of policies, with a
small delay in between. The test simply looks for no failures.