[SSPROD-5001] Send the compliance metrics through the events channel … (#2336)
* [SSPROD-5001] Send the compliance metrics through the events channel back to the
agent so they can be injected. This fixes a problem where the compliance metrics
are only reported for the host when the statsd forwarder is on.

Also, compliance metrics can no longer go through the syscall method because they
will get counted twice on the host after the fix for SMAGENT-2564. Since cointerface
was sending the metric directly to statsite (in the agent container), the metric
will get counted as a host-duplicate and a host-not-duplicate. Injecting through the
analyzer keeps this from happening.

* Update the compliance unit test.

* Rework after conversation with Mark Stemm

* Mark's comments