[SMAGENT-2817] disable secure actions in agentino (#2340)
trying to execute policy actions in the agentino is unsupported, and must be explicitly disabled, as there could be a rule with an action in the infra. This is done via a config set using the new "agentino" mode.

Other minor changes include:

adding a "direct" mode to allow agentino to talk directly to the BE...useful for testing
include the pdig binaries in the agentino container, again greatly eases testing
move agentino container back to alpine. busybox was just too restrictive
couple of other bug fixes, including rolling our own basic arg parsing since poco's seems to....not work...sometimes (not sure if it's optimization levels, musl, or some combination...)