Implement agentone <=> agentino communication layer (#2337)
* add first file

* outline

* agentino manager calls from network

* compilest

* two managers work!

* needs to compile message handlers

* need to do propagate_policies, send policies to new agentino, and test

* who knows

* intermediate

* ready for nathan

* [SMAGENT-2772] Ensure pod_restart_rate is always increasing

We had customers experiencing spikes in the pod_restart_rate metric.
This happens if pod_restart_count is reset to 0. Passing from the
actual value to 0 gives a negative rate that casted to an unsigned int
returns a  ~2^32 value. Considering that when a reset is done no
restart happens, we handle this corner case passing to the BE the max
between the calculated rate and 0

* Update to OpenSSL 1.1.1h and remove Python 2.6 (#2285)

* Update to OpenSSL 1.1.1h and remove Python 2.6
* Fix net test

* [SMAGENT-2663] make the agent build with either whole or split sysdig repo (#2232)

* Pull in Lua chisel files from sysdig repository

This removes one of the last dependencies on sysdig (not libsinsp/libscap).

* Add zlib includes to libscap build

This is only needed when using direct add_subdirectory with
a cmake-module enabled branch of libscap (it depends on having
zlib includes available via the cmake module, and since we don't
load the cmake module, the includes are missing).

This doesn't affect non-cmake-module build in any observable way.

* Actually support split repos in agent build

Either mount the sysdig source in /draios/sysdig (in the build container)
or the separate libsinsp/libscap repos (under /draios/libsinsp
and /draios/libscap, respectively).

* Don't build two tests with libscap-hayabusa

These reach into internals that are not exposed with the alternate
libscap implementation.

* remove redundant rsync in docker build entrypoint

Co-authored-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>

* [SSPROD-5535] Fixing netsec issues when audit stream is disabled (#2296)

This change allows netsec to work in the context of audit stream and
audit tap to be disabled. With this fix, the connection tinfo are
properly stored when netsec is enabled, and therefore allows netsec to
properly lookup container metadata.

Note that connections that are not related to a container are
discarded by netsec. This issue was causing netsec not to track and
report any connection, as no connections were actually being marked as
belonging to a container.

Signed-off-by: Lorenzo David <lorenzo.david@sysdig.com>

* SMAGENT-2783: Add component level file logging capability to sdjagent (#2293)

Manually tested with different log levels using configs like this:
~~~~~~~~~~~
log:
  file_priority: critical
  console_priority: critical
  file_priority_by_component:
    - "connection_manager: info"
    - "sdjagent: debug"
    - "analyzer: notice"
    - "someunknowncomponent: trace"
    - "dragent.configuration: someunknownlevel"
    - "someunparseablestring:blahblah"
    - "watchdog_runnable: info"
    - "sdchecks: notice"
~~~~~~~~~~~

* SMAGENT-2771 - fix zeros in envVariableHash (#2299)

* defer hash_environment to the emit phase for long-living threads

Signed-off-by: Matteo Bertrone <matteo.bertrone@sysdig.com>

* fix audit_tap.ut.cpp

Signed-off-by: Matteo Bertrone <matteo.bertrone@sysdig.com>

* [SMAGENT-2792] Remove obsolete config to automatically fill metric limit headroom (#2301)

* [SMAGENT-2794] Retry VM availability every x seconds

- reverts commit 3a424e9350d717789c7fae4899cc257aa4e23bbb.
- retry logic implementation:

if the very first attempt fails, the second attempt is executed after
a random period d where d ∈ [5, 30] in seconds
all successive attempt interval increase exponentially till reaching
max_availability_check_interval_sec.
max_availability_check_interval_sec default value is 10 minutes. It
can be changed with this property:

jmx:
   max_availability_check_interval_sec: 120

* Poll for error conditions as well as pending reads (#2295)

* SMAGENT-2795: Add component level file logging for cointerface (#2303)

* SMAGENT-2795: Add component level file logging for cointerface
Manually tested the change with variations of this config
~~~~~~
log:
  file_priority: critical
  console_priority: critical
  file_priority_by_component:
    - "connection_manager: info"
    - "sdjagent: info"
    - "analyzer: notice"
    - "someunknowncomponent: trace"
    - "dragent.configuration: someunknownlevel"
    - "someunparseablestring:blahblah"
    - "watchdog_runnable: info"
    - "sdchecks: debug"
    - "cointerface: debug"
~~~~~~~~~

* Francesco's comment: Added a perf enhancement to look up component
level file priority only once per subprocesses_logger process

* [SMAGENT-2659] Added listen & poll socket methods (#2267)

Added static methods to the cm_socket class that allow for a server listen loop and which 
provide a polling mechanism for cm_sockets.

* Checkpoint

* Agentino connection object

* Agentino message

* Checkpoint

* Post-rebase

* It's awful, but ready for push

* Cleaned up new connection just a bit

* Made handshake workflow fit existing code better

* Fix a lot of compiler issues

* Split off agentino_connection into .h/.cpp

* Handle handshake via callback

* A couple of working unit tests now

* Added an agentino_manager UT

* Fix valgind errors

* Added thread pool

Changed the processing of agentino events to be handled asynchronously
via thread pool.

* fix agentino-container manager interaction, properly unpack new
connection metadata

* Handle data messages from agentino

This is a big update that connects a lot of moving pieces together. The
biggest piece is obviously hooking up the poll_and_dispatch function
into the run loop of the agentino manager.

* put policies in handshake response

* fix UT failure

* Made agentino port configurable

* Added SSL toggle (might work)

* fix and validate locking

* make socket poll timeout configureable

* Fix overlap UT

* fix policy handling in fake agentino

* Handled error callback from listen loop

* add agentino handshake to connection manager

* Made FSM marginally more thread-safe

* Fixed some indentation issues in CM changes

* Misc TODOs and fix a couple of valgrind probs

* Unit test for disconnecting during handshake

* Change a few wonky test things

* add metadata handler to handshake

* add correct params to connection_manager for direct communication

* Fix duplicate entry in CMakeLists.txt

* make agentone build with agentino_manager as well

* agentone and agentino talk

* all runs together

* Added AM / CM integration tests

* add agentone build to container build job

* Fixed merge error

* Many UT fixes

* Thread pool

* review comments

* Fixed an occasional test failure

* review comments

* Fixed a few review comments

* Documented usage of thread pool

* review comments

* Fixed other UT

* refactor self ptr

* review comments

* Yet more UT fixes

* opject

* review comments

* Try number 29804983 to get the UTs to pass

* UTs

* agentino logging in fargate (#2368)

specify --debug-logging to enable it

Co-authored-by: Kevin Kauffman <speedyguy17@gmail.com>
Co-authored-by: Francesco Emmi <francesco.emmi@sysdig.com>
Co-authored-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>
Co-authored-by: lorenzo-david <44177297+lorenzo-david@users.noreply.github.com>
Co-authored-by: Naveen Bali <42355440+balinaveen@users.noreply.github.com>
Co-authored-by: Matteo Bertrone <matteo.bertrone@sysdig.com>
Co-authored-by: Bryan Seay <42394778+bryanseay@users.noreply.github.com>
Co-authored-by: Radu Andries <radu.andries@sysdig.com>