[SMAGENT-2408][SMAGENT-2834] precompute thread hashes (#2377)
Alright.....this turned into a huge change, but I'll try to spell it out, and if it comes down to it, I can break it down. but i'll try to annotate it

THe backend requires the PID field of the program message to be set to a hash that allows processes to be associated across time (even if they get a new pid). We perform this action in the aggregator. THere are two issues:

this is horrendously slow
we don't have the value a-priori to compute process uniqueness
The latter is a correctness problem, as the BE uses the hash exclusively for uniqueness, whereas we were using the raw fields, which doesn't have the proper sanitization done, namely

auto separator_loc = details.exe().find(": ");
hash +=
java_string_hash(details.exe(),
separator_loc == std::string::npos ? details.exe().size() : separator_loc);

That bit which ignores the string after the separator

So the fix for BOTH of these problems is thus:

compute the hash ONCE the first time we populate a protobuf with a process
cache the hash, and use it again, so it is never necessary to recompute for that process
store it in a new field in the protobuf that the aggregator can use for uniqueness
This sounded somewhat straightforward, but lead to a lot of issues, and a huge amount of the cleanup for such a small code change

the moving of the hash function to a new namespace
the tests all failed, because computation of the hash involves allocating a main thread ainfo, which involves accessing the machine info, which does not exist in test
reworking the thread_analyzer_info to use cpu directly, instead of trying to access the machine info, thus allowing trivial injection of that dependency
modifying a bunch of function constness to get it to compile
reworking the threadinfo in sinsp to allow me to be able to comprehend what was going on.
So that said, I will annotate this review, and here is the relevant libsinsp review:
draios/libsinsp#3
draios/sysdig#1734
draios/protorepo#105

Recommend do the libsinsp review first, and read my annotations as you go along to hopefully explain that there is very little actual substantial change here.