SSPROD-6330 intuit missing policy type (#2403)
* If policy type is missing, do syscall+k8s_audit

When a policy does not have a policy type at all, add the policy to both
the syscall rules group and the k8s audit rules group. This ensures that
k8s audit rules within policies that do not have a type can still match
k8s audit events.

This is important for existing policies that do not have a type, or for
older backends that have not been upgraded to add a policy type to
policies.

This fixes SSPROD-6330.

* Test for policies + no policy type + k8s audit

A new test has a policy with no policy type and with rules that work on
k8s audit events. The test verifies that the policies still result in
policy events when matched against k8s audit events.