add hostname to machine_id_prefix in fargate (#2407)
mac address is insufficient for uniqueness in fargate. we already require customers to pass in a hostname if they are talking to the backend (agentone=always, agentino=if using direct). As such, we will require customers to use a unique value for this and stuff it in the machine_id_prefix.

We will need documentation that hostname must be unique per customer.