[SSPROD-6724][SSPROD-6113] fargate release job and fix build_* scripts
There are a few changes here, noted by changeset:

fixup fargate release ports two changes that were used during fargate release as well as adds missing steps to the fargate release job, which hopefully we can use for v2. The biggest missing part is we need a job to invoke on the kilt side.
fix scan_container+fix_artifactory_login: this was YOLO broken sometime in january. I fixed it and added support for artifactory-based containers
The rest: the upload-agent + friends bash script has been a thorn for quite some time, since it did three logically distinct things:
upload our packages to s3
build production containers
upload containers
This was further complicated by the fact that this script had 3 flavors each which demanded slightly differing behavior:
build_dev
build_rc
build_farate_rc

So the logical thing to do was to break the script up into 3 parts, each parameterized appropriately so it can be shared between the 3 jobs, so that's what I did

upload-agent now uploads packages to s3
build-prod-container now builds any or all of the prod containers from packages in whatever s3 bucket you tell it
the individual then can tag and upload those containers wherever it wants.

The upload-agent script is untouched, and was used in the previous fargate release, so that "should" work. the build-prod-container script was tested locally against both the single-image variant and all-image variants, so that should work. There is not a good way to do a dry run of build_dev/build_rc, so those will simply have to be monitored during the next run for any issues.

The last couple of thorns in my side which will not be addressed here are:

the promote-rc script, which duplicates a lot of the work of upload-agent, including dealing with the RC buckets, and most especially, rebuilding the docker containers. this is where the "stable" tag is built. This will likely break agentone, which doesn't (though should) have a stable tag.
the build_dev and build_rc jobs ought be merged.