Handle cont. switches for pids in sec. policies (#2503)
Prevent using the wrong policies group for processes that change
containers. This is most common for container execs, where a process
starts on the host, switches to a container, and then might exec to the
desired exec.

The bug only occurs for kube execs/docker execs where for a given pid, its
container context changes from the host to into a container, and where
all the syscalls the security manager sees between the clone/exec that
creates the process, the clone (changing into the container context) and
the exec (launching the process) all are for the same pid.

Also, a given rule (in this case Terminal Shell in Container) must be
used by multiple policies at different scopes.

The security manager code has this optimization where the security rules
group for a pid uses the last value if the pids are the same:

---
if(tinfo->m_pid != m_last_pid)
{
    m_last_security_rules_group =
        m_loaded_policies->get_rules_group_for_container(*container_id_ptr);
}
m_last_pid = tinfo->m_pid;
---

What's happening in this case is that a given pid starts out in host
context. At that time, the security rules group is the group for the
host (container.id = ""). Then the clone changes the container context
to the container. But since the pid hasn't changed, the security rules
group isn't refetched. Finally, the exec occurs and the wrong set of
falco rules are evaluated--the ones for the host and not the ones for
the container.

The fix is to consider both the pid and container id when deciding
whether to look up the rules group again.