Handle local changes when syncing w upstream falco (#2502)
We now have local changes in draios/oss-falco that aren't in
falcosecurity/falco, so we can't just take falcosecurity/falco and put
it on a candidate branch for testing.

Change how we prepare the local sync branch to use rebasing instead of
merging, following the approach in
https://github.com/draios/agent/pull/2464, which does a similar job of
syncing a fork of falcosecurity/libs.

The "sync branch" isn't master as that had been used for merges and I
don't want to modify it. We'll use master-sync instead.

If this works, I'll change the agent build to use master-sync instead of
master.