[SMAGENT-2934] Log exception when statsd fails to start (#2494)
Log the exception if thrown when calling the nsenter constructor. Also, add the errno to the exceptions returned from the nsenter constructor as it will provide useful information into what is going wrong if it fails. If we catch an exception while trying to initialize the statsd_server, log the exception.

There were no pre-existing unit tests for the statsite_forwarder so I manually tested the modified log by adding a thrown exception in the try block and verifying it was logged. nsenter is tested via agent/test/container.cpp and I've modified one of the tests to verify that errno is included in the exception.

I noticed that we had a .clang-format file so I ran clang-format on the files I changed. That is the cause for the various formatting changes in the files unrelated to the other changes.