[SSPROD-6480] Forward THROTTLED_POLICY_EVENTS (#2483)
Sometimes the security manager will determine that the agent is being
too noisy with policy events and will begin to throttle them. In a
Fargate context, this means that the agentino will send a
THROTTLED_POLICY_EVENTS message to the agentone instead of a
POLICY_EVENTS message. The agentone should forward this message to the
backend for feature parity with the stock agent.

(That said, I have been informed that the backend does nothing with
 these THROTTLED_POLICY_EVENTS messages, and thus we are not bending
 over backward to implement this functionality.)