Switch to master-sync to build draios/oss-falco (#2507)
* Switch to master-sync to build draios/oss-falco

Now that the jenkins job falco-test-and-sync-upstream makes changes to a
master-sync branch, we need to update builds that use draios/oss-falco
to use the master-sync branch as well.

Update all jenkinsfiles to use master-sync instead of master to perform
builds.

We'll also need to change the default value for any parameters in
jenkins jobs to use master-sync as well. We can do that right after
merging these changes.

* Change presubmit.jenkins to use master-sync.

* Use semver for tags on master-sync

It turns out that other parts of the falco build look at the most recent
tag to set the MAJOR/MINOR/PATCH versions, so make our tags have a
major.minor.patch, where we use the jenkins build as the patch.