[SMAGENT-2984] gzip agentino handshake with BE (#2528)
we were missing the proper initialization for the compressor factory. I (hopefully) simplified it to make ti easier for future applications to use, as well as made it explicit that the handshakes use gzip, since the protocol demands it. If any of the changes are ill-conceived or otherwise broken, I'm happy to revert parts of it, as simply adding the two lines that were in agentone and dragent would also have solved the problem. But I thought this might help prevent it 🤷

We also move to actually flush every 10s. In theory this dummy flush should be shared code between the agentone and agentino....but it is not.