[SSPROD-6937] Delay startup until policies loaded (#2510)
There is a window where the workload is unsecured because the policies
haven't been loaded yet. We can work around this gap by not starting the
instrumented workload until the policies have been received and loaded
successfully. The tradeoff with using this option is that if the
agentino can't connect to the agentone or if the agentone can't connect
to the backend, the workload will not start at all.

As we are a security-first company and we don't want to leave our
customers vulnerable out of the box, we have chosen to turn this option
on as the default.