[SMAGENT-3016] UBI7 based builder
Our choices for the base image are limited, since ubi8 does not
support the scl repo (AFAICT).

ubi7-minimal only has microdnf instead of yum, which cannot download
packages as rpm files, just install them, so we work around this
limitation with a custom script.

We need the download functionality because of course UBI is
incompatible with CentOS packaging so we cannot cleanly install
e.g. dkms.

Also, we apparently we need `microdnf clean all`
between package installations.

We follow the same approach as in the Debian builder, symlinking
various gcc versions, but in this image, they all come from
either the base repo (gcc 4.8, masquerading as 5.0, for the same
reason as in the Debian builder), or from devtoolset-* packages.

Since the default binutils version is ancient on UBI7, we don't
need to worry about downgrading them.

To pass our security image scans, we remove a bunch of unneeded
packages (though yum and systemd both have a ton of dependencies
and cannot be uninstalled, so some cruft remains) and strip
set*id bits from all files.

Some notable hacks include:
* rewriting the docker-ce repo file to replace the `7Server` release
  with just `7`. See https://github.com/docker/for-linux/issues/1111

  Note, yum errors out because it cannot follow a 301 redirect
  that does lead to an existing file. Whether that is solvable within
  yum (e.g. by enabling https support?), I did not investigate

* LD_LIBRARY_PATH=/opt/draios/lib is needed for the `postgres` app check.
  While UBI does ship with libpq.so.5 (in `postgresql-libs`), that
  version is not ABI-compatible, apparently (missing symbols).

* There's no python3 package available that would install `/usr/bin/python3`.
  So install `rh-python38-python` and link the interpreter
  in the right place.

  Note: `/usr/bin/python3` is so that you can just run `/opt/draios/bin/sdchecks`
  and `/usr/bin/python3.8` is the preferred Python interpreter name
  within dragent.

* We use the systemwide Mozilla CA bundle at /etc/pki/tls/certs/ca-bundle.crt.
  `ca-certificates` is already installed, so no need to install
  the system Python certifi package.

* `iproute` is needed for some benchmark tool called from cointerface
  (it uses `ss` provided by this package)