[ESC-857][SMAGENT-3011] add deadman to running state terminator (#2540)
There are cases where the main dragent thread can hang (especially during startup, and especially in sinsp work init...these are problems unto themselves and are not addressed here). This adds a deadman on restart so worst case we kill the process instead of hanging. It also prints a stack trace of the main thread, which is likely where we're stuck if shutdown does not complete.