heavy handed cleanup to jenkins, and fix agent libs sync job (#2547)
two functional changes here:

sync_agent_libs was not using the appropriate non-legacy libs. this recuired changes to the job to do the sync as well as enhancing the two presubmit test jobs. The corresponding configs were added to the presub configurations
i'm tired of cleaning up jenkins boxes.....so we're going total tank on cleanunp. removing everythign in the workspace before marking it as "cleanup" I tested this, and it works, leavingonly a directory behind. I don't believe anyone has any use for the "cleaned up" directories.