[SSPROD-7590] infra-state agnostic container scope matching (#2554)
if infra state does not exist, no reason to depend on infra state. instead use container manager directly as the metadata source. Infra state matching is left 100% in tact, so should have no impact on agent, even if we could use direct container matching in the regular agent.

I intended to make the format commit, which brings security_rule.cpp to the rest of the agent coding standard a separate commit, but for some reason that appears to be not doing as i indended. I can try again if people woul dlike. In any case, all functional change in that file is contained to match_syscall_scope.

Dependent libsinsp review: draios/libsinsp#10