[SMAGENT-3027] Add and update missing fields for unified workload metrics (#2569)
These changes add some metrics that are needed for unified workload metrics. There is a table that maps Kubernetes API metrics to unified workload metrics in the product spec.

The metrics added are status.numberAvailable, status.numberUnavailable, and status.updatedNumberScheduled for daemon sets, status.availableReplicas and status.replicas for replica sets, and status.replicas for deployments.

To test, I modified and ran the aggregator unit test as well as all of the others. I also deployed on a k8s cluster, dumped the protobufs, and checked that the new metrics were present and the values were correct.