[SMAGENT-3065] eliminate sample handler deadman (#2609)
the sample handler is a "helper" of the serializer which deals with flush data. the serializer itself has its own thread and has a deadman which fires if that thread does not complete a loop. It wakes up far more often than the deadman, so it should be the case that THAT deadman only fires when there is work to do and it is not being done.

The sample handler ALSO had a deadman associated with it. This deadman would fire if we hadn't emitted a sample for a specified amount of time. Just because we haven't emitted a sample, however, does not mean we are not making progress, or that there is even work to do. This is even more relevant now that we are pushing to higher and higher flush intervals.

As such, this deadman was not providing much real value, and killing the agent, we end up more likely to not recover than if we let the agent sort itself out. So we'll let the numerous other stall detection methods (like, erherm, sinsp worker) deal with actual stalls and slowness, and we will eliminate this one. We will, however, leave the infra in place and just put a rate-limited log message.