[SSPROD-7548] Support Linux versions which eliminate deprecated SYS_OPEN system call
Recent versions of Linux have dropped support for the deprecated
SYS_OPEN system call, in favor of SYS_OPENAT which provides a
superset of the SYS_OPEN functionality.  Enhance test code which
directly generates system calls (instead of using the libc wrappers),
to use SYS_OPENAT when SYS_OPEN is not available.