[SSPROD-8254] create new centos builder container (#2633)
When the alpine builder was created, we standardized how we build dependencies into the container. This change ports that to the centos builder, including some other improvements

streamlining of entrypoint commands
not downloading prom on each invocation (no way to force reinstall yet...)
on-demand invocations of cmake....only run what is needed
elimination of bootstrap-agent script which runs every time. dependencies are fixed at container build time
output directories are named after the build variant in them, instead of a loosely similar name with different capitalization and delimiting
combining of scripts to install dependencies between alpine and centos builders (could be much more done here to make it dynamic...but baby steps)
This change will go in for some guinea pig testing before we consider rolling it out to production, as some changes need to happen to jenkins (to deal with the newly named directories and slightly more involved script to build it). A candidate for testing is located at: quay.io/sysdig/agent-builder:new-beta

Minor change to the build-alpine jenkins job to deal with this.