[SMAGENT-3108] allow consistent creation of existing cpu cgroup
in order to be able to save the existing cpu cgroup,
(so to restore it later), we need to have a fully functional
subprocess_cpu_cgroup object for an already-existing cgroup,
which can be easily created by having a subprocess_cpu_cgroup
with empty name and 0 values.
For this to work though, we need m_created to be set to true.
The overriding subprocess_cpu_cgroup::create() method will not call
the overridden method in such cases, so m_created will never be updated.

So initialize m_created according to whether the cgroups path already
exists or not.