[SMAGENT-3108] Add monitor process to default cpu cgroup
All created processes are moved to the default cpu cgroup.
The creator however, stays with the original cgroup.
While this should not normally be an issue (since it its job
is just to fire up all subprocesses and just monitor them
at 1-s intervals), there might be corner cases where it is
(e.g. when the agent keeps crashing at startup for whatever
reason).
So move it to the default cgroup as early as possible in order
to mitigate such effects.

Also, move it back to its previous cgroup during cleanup,
so that if the agent gets restarted, there are no side effects.

(WARNING: requires the cgroup object to have a consistent
m_created field)

Tested by commenting the following line on monitor.cpp:258

    //this_thread::sleep_for(chrono::seconds(1));

and manually verifying how with this change, the CPU usage
of the main process is indeed limited to 25%
(see SMAGENT-3108 for details).

Co-authored-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>