Fix inverted logic when checking whether image name is the id
This got broken when adding Podman support. The original code involved
multiple layers negations and while trying to understand the code,
I got the logic wrong. This got exposed when running a container
using the image id instead of the image name: we mistakenly reported
the image repo/tag as sha256/<image id>.