[SMAGENT-3056] Add init container statuses to k8s_pod (#2678)
This change adds init containers to the k8s_pod but only if the config option to send all containers is set. The reason for this is that we didn't send init containers at all previously so we wanted to preserve that behavior unless the new config is enabled.

Added to unit tests and tested in k8s cluster manually.