[SMAGENT-3164] Emit warning upon cgroup creation failure
When running the agent in a container managed by kubernetes,
cpu limits should be specified by means of kubernetes'
resource requests/limits mechanisms.
However, customers used to running the agent outside of
an orchestrated environment, where they can limit the agent's
resources by means of configurations parameters, may still
be tempted to use the old cgroup-based approach and expect
it to work.
While in theory this approach should work in a containerized
environment, in practice the host's cgroups are normally
mounted read-only under /host/cgroup/{subsystem} and are
therefore not easily accessible, so this apporoach will fail.

If that is the case, add a warning message in order not to
create false expectations.