Build agent-slim-ubi images with UBI8
We don't need any build tools in the slim image so we can use UBI8 there.

Note: the new template is as similar as possible to the UBI7 one,
even though we can't build a running non-slim image on UBI8. This
is intentional to simplify maintenance of the two Dockerfiles.

In the future we might be able to build a working agent/kmodule container
on UBI8, e.g. by shipping our own build tools.