[SMAGENT-2998] Drop usage of default self-signed CA certificate
SSL connection to the collector backend currently includes both a
hardcoded CA with a self-signed certificate and the path where CA
certificates are stored.

Remove default usage of the sysdig-provided self-signed certificate,
while keeping the option for the user to provide their own.
In other words, change the default value for the config option
to an empty string and make sure this case is handled correctly.