[SMAGENT-3184] Don't send KSM container metrics in local_kubernetes (#2725)
To decrease the amount of data being sent and stored in the backend, we will stop sending the new container-level KSM metrics in local_kubernetes. We will continue to send the same containers (waiting and terminated) that we sent previously in local_kubernetes.

This change also adds the ability to export differently depending on if the caller is exporting globally or locally to the legacy_k8s_exporter.

Tested by running modified unit tests.