[QA-2943] Fix agent-kmodule UBI image
The kmodule UBI image only ships with the draios-agent-kmodule
package installed (no draios-agent-slim) so it's missing
the sysdigcloud-probe-loader script.

Once we add it, we also need to explicitly specify the agent version
(using a script that checks the installed rpm version), as otherwise
the script attempts to run dragent (which isn't there either).