fix build_dev and build_rc auth to docker.internal.sysdig.com
build_dev and build_rc will sudo-call build-prod-container script.
This script may need access to our internal artifactory to build images.
Since it's called within a docker.withContainer() block, it may not inherit
the authentication provided by the latter.
Use sudo -E to make it work.

Note: it is not clear how this could have ever worked in the past.
Most probably it was due to accidentally cached credentials or images.