[SMAGENT-2856] generalize connection_server (#2718)
the goal of this change is to extract the networky bits from agentino_manager to create a reusable connection_server for anyone who needs to talk to people. This leaves the agentino_manager to just do agentino specific things.

From a high level, we create a connection_server, and the user of that (in this case the agentino manager) becomes a connection_server_owner, which entails an API of virtual functions that were previously implemented as c-style callbacks.

The change is large, and broken down nicely into commits for your conveninece. Let me run them down:

rename and reformat: as named. formats some of these files which weren't, moves agentino_connection->connection. no functional changes

create connection server and direct move functions from agentino manager: the connection_server.h file is created, and much of the code that we will leverage is DIRECTLY MOVED from agentinoi_manager->connection_server. The functions in connection_server.cpp which are still scoped to agentino_manager are DIRECT COPIES. ones scoped to connection_server are new code.

modify connecion_server functions to deal with not being in agentino manager anymore: the functions which were moved in the previous commit are modified to work in their new location. it should be relatively clear how the old code maps to the same functionality in the new

cleanup connectino with new changes, move handshake code to agentino_manager unmodified: the handshake code is copied directly from connection->agentino_manager. and various functions are changed to deal with the new signatures in the connection_server. creation of the connection_context to replace the "handshake_data"

fix agentino_manager: make the changes to make the agentino an actually correct user of the connection_server. the biggest changes are using the APIs instead of callbacks, and the "correction" of the handshake method

make source compile: various changes to make stuff compile. should be no functional change, other than removing the listen/stop_listening functions which should have been removed in an eralier commit but weren't

uts: move the connection_server specific tests to a new file, and change the tests so they compile now

finish completing unit tests: actually make the unit tests pass

dependentant protorepo review: draios/protorepo#149