[SSPROD-8258] Fix calling of build_gs downstream job
agent-build-gs is one of those jobs whose parameters
defaults (i.e. branch names) are all calculated as a
function of the base release.
Having them resolved correctly proved more difficult
and confusing than expected, so it's easier to
populate them correctly from the upstream job instead.

NOTE: The downstream job build_gs.jenkins should
also be adapted so to remove such broken default values.
We'll do it on a separate patch in order to keep this
patch to the bare minimim and therefore reduce
conflicts on cherry-picking.