[SMAGENT-2878] Part 1: move compliance manager cruft from sinsp_worker (#2746)
the ultimate goal of this task (2878) is to pare apart sinsp_worker to get independence of secure and the analyzer (and compliance manager and capture job handler) to allow the infras to be completely independent, allowing them to run without the others trivially or in parallel.

In order to kill sinsp_worker, though, we have to start moving much of the things that have sort of been orphaned there over the years to where they belong. This change aims to do that with compliance manager, specifically the infra around handling the infra around initialization and message handling (actual runtime will stay there for now. This change is broken up into three changes which are described below:

move compliance_statsd_destination from sinsp_worker to analyzer: This is the easiest. previously, we send the compliance statsd outgoing metrics on a loop-de-loop through the sinsp_worker. Now we send them directly to the agent, allowing the agent to handle them. While we likely need a better way (in general) to handle remote infras with statsd metrics to report, there is no value to going through the sinsp_worker.

move security/compliance initialization to dragent: while the second task WAS to move the incoming message handlers to the compliance manager directly, but this posed a problem, as we had this lazy initialization with both the secure and compliance infra, and that lock. This is the problem when lazy initialization is not owned by the infra itself, every other user of that infra must suddenly be aware that it is being lazily initialized. Given the model we will move to with neo-config, we cannot justify a special-case lazy initialization at this level, and we should move to these being initialized as the rest of the components in the agent are. If the module itself would like to delay some of its initialization WITHIN the module, it should be free to, but that is and should be opaque to any users of the module's api.

In this case, this means we initialize, and may restart when the config arrives. This is only an issue the first time the agent starts ever, and given we don't have policies, it isn't doing much work to begin with. If we actually depend on the lazy initialization, we can build that inside security_mgr and compliance_mgr, not leak that implementation.

Moving them to the normal initialization are means they are guaranteed to be set up before anyone uses them. This avoids any of the mutex stuff and checking whether they are initialized yet. They are then passed directly into the sinsp_worker.

Unlike the first change, this one does not compile and requires the third change to be a complete thing

move the handlers for calendar and task runner out of sinsp worker into compliance manager: this is what it sounds. we move the two handlers from sinsp_worker directly into compliance manager. Due to the previous commit, it is guaranteed to be initialized an ready to receive those messages, as we delay connection_manager startup until after we have registered the handler (and there is no more lazy initialization). The infra to support caching is simply deleted, and the change is complete.